set stream everytime when we get a cuBlas handle (#31537)
Summary:
I don't see any reason for not doing so, because it is a common error that people forget to set the stream. And I don't think there is a reason for not running on the current stream.

This is just for cublas, cusparse and cudnn should be modified also.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31537

Differential Revision: D19206908

Pulled By: ngimel

fbshipit-source-id: ba2b2b74e9847f0495c76dbc778751a9f23f8b36