bad tbb lambda capture, bad chunk size (#30352)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30352

1) tbb forwards us ident through parameter, we don't need to capture it.
2) tbb is being passed steps <= 0 which is bad.

Taken from TBB documentation:
```
The index type must be an integral type. The loop must not wrap around. The step value must be positive. If omitted, it is implicitly 1.
```

I have a build that uses `TBB_USE_DEBUG=1` and there are currently a lot of issues with PyTorch use.
Is TBB version not tested very much right now?
ghstack-source-id: 94459382

Test Plan: CI green

Differential Revision: D18666029

fbshipit-source-id: d5aa8327b03181d349e1964f9c8211298c433d6a