[clang-format] Replace asyncio.run with approximation supported in python 3.6 (#35501)
Summary:
**Summary**
`asyncio.run` is supported only after 3.7 and that too, provisionally so.
This commit replaces the use of `asyncio.run` in `tools/clang_format.py`
with an approximation that works in both 3.6 and 3.7.

**Testing**
Ran the script with both `python3.6` and `python3.7`.

```
$ python3.6 tools/clang_format.py --diff
...
Some files not formatted correctly
$
```

```
$ python3.7 tools/clang_format.py --diff
...
Some files not formatted correctly
$
```
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35501

Differential Revision: D20681947

Pulled By: SplitInfinity

fbshipit-source-id: 43e13aa85f79396bec1f12ee1e80eff90dbed5db