CI failure tips (#36507)
Summary:
Finding out how to ssh into a CircleCI job to debug a failure is a challenge because, as far as I know, there isn't any concise documentation about it. I figured it might be nice to include this in CONTRIBUTING.md.

Maybe there are some other tips about non-CircleCI jobs that could be added in the future as well.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36507

Differential Revision: D21006526

Pulled By: ezyang

fbshipit-source-id: 0a544ecf37bf9550e9b2f07595332dc5f394bb9e