Unconditionally register schema even for manual registration. (#36250)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36250

The general concept is that I want a centralized location where you
can find all of the registrations for a library.  I cannot do this
if I don't codegen all of the schemas in one spot--right now,
most schemas get generated, but not manually registered ones. Let us
assume that manual registration has to do with the actual
implementations; nothing strange is going on with the schema
definition itself.  Make it so.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

Test Plan: Imported from OSS

Differential Revision: D20929258

Pulled By: ezyang

fbshipit-source-id: 0a9fdc8eccd7b688b3e7bd8ed64b6e2af21978f4